backend/services/telegram_bot.py
"""
Telegram Bot Service (Optional)
"""

import os
from telegram import Bot


class TelegramBot:
    def __init__(self):
        self.token = os.getenv('TELEGRAM_BOT_TOKEN')
        self.chat_id = os.getenv('TELEGRAM_CHAT_ID')
        if self.token:
            self.bot = Bot(token=self.token)
        else:
            self.bot = None
    
    async def send_message(self, message: str, chat_id: str = None):
        if not self.bot:
            return
        target_chat = chat_id or self.chat_id
        if target_chat:
            await self.bot.send_message(chat_id=target_chat, text=message)
    
    async def send_reminder(self, reminder: dict):
        message = f"⏰ Reminder: {reminder['message']}"
        await self.send_message(message)
    
    async def handle_update(self, update: dict):
        # Handle Telegram webhook updates
        # This is a placeholder - full implementation would route to main.py endpoints
        pass

