backend/requirements.txt
# RAVVYN AI Assistant - Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# AI Providers
openai>=1.0.0
google-generativeai>=0.3.0  # For Gemini API

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0

# Telegram Bot (optional)
python-telegram-bot==20.7

# PDF (optional)
PyPDF2==3.0.1

# Date/Time
python-dateutil==2.8.2

# Scheduler for background tasks
apscheduler==3.10.4

# Note: CORS is built into FastAPI, no separate package needed

